Implement email validator in go
Added a new 'Validator' package in Go for verifying email formats. This package includes, necessary functions to check if the value of the email field in the Validator struct matches the email Regular Expression and if the struct has no errors. Additionally, it provides functions for validating permitted values and uniqueness.